Add integration test for harbor webhook
This surfaced a bug in how we defined referenced the pointer to the
HarborWebhook struct during demarshaling.

For now we just check that the response was good, but in the future it'd
be cool to rely on the kubernetes client lib a bit more to check that it
actually made a patch to the deployment as expected.